make the tests run with all runners
