+++ db-service/test/cqn4sql/helpers/expectCqn.js
@@ -1,0 +1,30 @@
+'use strict'
+
+const cds = require('@sap/cds')
+const { expect } = cds.test
+
+const normalize = q => JSON.parse(JSON.stringify(q))
+
+/**
+ * Usage:
+ *   expectCqn(actual).to.equal(expected)
+ *   expectCqn(actual).not.to.equal(expected)
+ */
+function expectCqn(actual) {
+  const a = normalize(actual)
+  const chain = expect(a)
+  return {
+    to: {
+      equal: expected => chain.deep.equal(normalize(expected)),
+    },
+    get not () {
+      return {
+        to: {
+          equal: expected => chain.not.deep.equal(normalize(expected)),
+        }
+      }
+    }
+  }
+}
+
+module.exports.expectCqn = expectCqn
